test: split up naming-convention
