+++ packages/eslint-plugin/tests/rules/naming-convention/cases/accessor.test.ts
@@ -1,0 +1,17 @@
+import { createTestCases } from './createTestCases';
+
+createTestCases([
+  {
+    code: [
+      'const ignored = { get %() {} };',
+      'const ignored = { set "%"(ignored) {} };',
+      'class Ignored { private get %() {} }',
+      'class Ignored { private set "%"(ignored) {} }',
+      'class Ignored { private static get %() {} }',
+      'class Ignored { static get #%() {} }',
+    ],
+    options: {
+      selector: 'accessor',
+    },
+  },
+]);
