+++ packages/eslint-plugin/tests/rules/naming-convention/cases/class.test.ts
@@ -1,0 +1,10 @@
+import { createTestCases } from './createTestCases';
+
+createTestCases([
+  {
+    code: ['class % {}', 'abstract class % {}', 'const ignored = class % {}'],
+    options: {
+      selector: 'class',
+    },
+  },
+]);
