+++ packages/eslint-plugin/tests/rules/naming-convention/cases/enum.test.ts
@@ -1,0 +1,10 @@
+import { createTestCases } from './createTestCases';
+
+createTestCases([
+  {
+    code: ['enum % {}'],
+    options: {
+      selector: 'enum',
+    },
+  },
+]);
