+++ packages/eslint-plugin/tests/rules/naming-convention/cases/interface.test.ts
@@ -1,0 +1,10 @@
+import { createTestCases } from './createTestCases';
+
+createTestCases([
+  {
+    code: ['interface % {}'],
+    options: {
+      selector: 'interface',
+    },
+  },
+]);
