+++ packages/eslint-plugin/tests/rules/naming-convention/cases/parameter.test.ts
@@ -1,0 +1,21 @@
+import { createTestCases } from './createTestCases';
+
+createTestCases([
+  {
+    code: [
+      'function ignored(%) {}',
+      '(function (%) {});',
+      'declare function ignored(%);',
+      'function ignored({%}) {}',
+      'function ignored(...%) {}',
+      'function ignored({% = 1}) {}',
+      'function ignored({...%}) {}',
+      'function ignored([%]) {}',
+      'function ignored([% = 1]) {}',
+      'function ignored([...%]) {}',
+    ],
+    options: {
+      selector: 'parameter',
+    },
+  },
+]);
