+++ packages/eslint-plugin/tests/rules/naming-convention/cases/typeAlias.test.ts
@@ -1,0 +1,10 @@
+import { createTestCases } from './createTestCases';
+
+createTestCases([
+  {
+    code: ['type % = {};', 'type % = 1;'],
+    options: {
+      selector: 'typeAlias',
+    },
+  },
+]);
