+++ packages/eslint-plugin/tests/rules/naming-convention/naming-convention.test.ts
@@ -1,12 +1,6 @@
 /* eslint-disable @typescript-eslint/internal/prefer-ast-types-enum */
-import { TSESLint } from '@typescript-eslint/utils';
-import rule, { MessageIds, Options } from '../../src/rules/naming-convention';
-import {
-  PredefinedFormatsString,
-  Selector,
-  selectorTypeToMessageString,
-} from '../../src/rules/naming-convention-utils';
-import { getFixturesRootDir, RuleTester } from '../RuleTester';
+import rule from '../../../src/rules/naming-convention';
+import { getFixturesRootDir, RuleTester } from '../../RuleTester';
 
 const ruleTester = new RuleTester({
   parser: '@typescript-eslint/parser',
@@ -19,717 +13,8 @@ const parserOptions = {
   project: './tsconfig.json',
 };
 
-const formatTestNames: Readonly<
-  Record<PredefinedFormatsString, Record<'valid' | 'invalid', string[]>>
-> = {
-  camelCase: {
-    valid: ['strictCamelCase', 'lower', 'camelCaseUNSTRICT'],
-    invalid: ['snake_case', 'UPPER_CASE', 'UPPER', 'StrictPascalCase'],
-  },
-  strictCamelCase: {
-    valid: ['strictCamelCase', 'lower'],
-    invalid: [
-      'snake_case',
-      'UPPER_CASE',
-      'UPPER',
-      'StrictPascalCase',
-      'camelCaseUNSTRICT',
-    ],
-  },
-  PascalCase: {
-    valid: [
-      'StrictPascalCase',
-      'Pascal',
-      'I18n',
-      'PascalCaseUNSTRICT',
-      'UPPER',
-    ],
-    invalid: ['snake_case', 'UPPER_CASE', 'strictCamelCase'],
-  },
-  StrictPascalCase: {
-    valid: ['StrictPascalCase', 'Pascal', 'I18n'],
-    invalid: [
-      'snake_case',
-      'UPPER_CASE',
-      'UPPER',
-      'strictCamelCase',
-      'PascalCaseUNSTRICT',
-    ],
-  },
-  UPPER_CASE: {
-    valid: ['UPPER_CASE', 'UPPER'],
-    invalid: [
-      'lower',
-      'snake_case',
-      'SNAKE_case_UNSTRICT',
-      'strictCamelCase',
-      'StrictPascalCase',
-    ],
-  },
-  snake_case: {
-    valid: ['snake_case', 'lower'],
-    invalid: [
-      'UPPER_CASE',
-      'SNAKE_case_UNSTRICT',
-      'strictCamelCase',
-      'StrictPascalCase',
-    ],
-  },
-};
-
-const REPLACE_REGEX = /%/g;
-// filter to not match `[iI]gnored`
-const IGNORED_FILTER = {
-  match: false,
-  regex: /.gnored/.source,
-};
-
-type Cases = {
-  code: string[];
-  options: Omit<Options[0], 'format'>;
-}[];
-function createValidTestCases(cases: Cases): TSESLint.ValidTestCase<Options>[] {
-  const newCases: TSESLint.ValidTestCase<Options>[] = [];
-
-  for (const test of cases) {
-    for (const [formatLoose, names] of Object.entries(formatTestNames)) {
-      const format = [formatLoose as PredefinedFormatsString];
-      for (const name of names.valid) {
-        const createCase = (
-          preparedName: string,
-          options: Selector,
-        ): TSESLint.ValidTestCase<Options> => ({
-          options: [
-            {
-              ...options,
-              filter: IGNORED_FILTER,
-            },
-          ],
-          code: `// ${JSON.stringify(options)}\n${test.code
-            .map(code => code.replace(REPLACE_REGEX, preparedName))
-            .join('\n')}`,
-        });
-
-        newCases.push(
-          createCase(name, {
-            ...test.options,
-            format,
-          }),
-
-          // leadingUnderscore
-          createCase(name, {
-            ...test.options,
-            format,
-            leadingUnderscore: 'forbid',
-          }),
-          createCase(`_${name}`, {
-            ...test.options,
-            format,
-            leadingUnderscore: 'require',
-          }),
-          createCase(`__${name}`, {
-            ...test.options,
-            format,
-            leadingUnderscore: 'requireDouble',
-          }),
-          createCase(`_${name}`, {
-            ...test.options,
-            format,
-            leadingUnderscore: 'allow',
-          }),
-          createCase(name, {
-            ...test.options,
-            format,
-            leadingUnderscore: 'allow',
-          }),
-          createCase(`__${name}`, {
-            ...test.options,
-            format,
-            leadingUnderscore: 'allowDouble',
-          }),
-          createCase(name, {
-            ...test.options,
-            format,
-            leadingUnderscore: 'allowDouble',
-          }),
-          createCase(`_${name}`, {
-            ...test.options,
-            format,
-            leadingUnderscore: 'allowSingleOrDouble',
-          }),
-          createCase(name, {
-            ...test.options,
-            format,
-            leadingUnderscore: 'allowSingleOrDouble',
-          }),
-          createCase(`__${name}`, {
-            ...test.options,
-            format,
-            leadingUnderscore: 'allowSingleOrDouble',
-          }),
-          createCase(name, {
-            ...test.options,
-            format,
-            leadingUnderscore: 'allowSingleOrDouble',
-          }),
-
-          // trailingUnderscore
-          createCase(name, {
-            ...test.options,
-            format,
-            trailingUnderscore: 'forbid',
-          }),
-          createCase(`${name}_`, {
-            ...test.options,
-            format,
-            trailingUnderscore: 'require',
-          }),
-          createCase(`${name}__`, {
-            ...test.options,
-            format,
-            trailingUnderscore: 'requireDouble',
-          }),
-          createCase(`${name}_`, {
-            ...test.options,
-            format,
-            trailingUnderscore: 'allow',
-          }),
-          createCase(name, {
-            ...test.options,
-            format,
-            trailingUnderscore: 'allow',
-          }),
-          createCase(`${name}__`, {
-            ...test.options,
-            format,
-            trailingUnderscore: 'allowDouble',
-          }),
-          createCase(name, {
-            ...test.options,
-            format,
-            trailingUnderscore: 'allowDouble',
-          }),
-          createCase(`${name}_`, {
-            ...test.options,
-            format,
-            trailingUnderscore: 'allowSingleOrDouble',
-          }),
-          createCase(name, {
-            ...test.options,
-            format,
-            trailingUnderscore: 'allowSingleOrDouble',
-          }),
-          createCase(`${name}__`, {
-            ...test.options,
-            format,
-            trailingUnderscore: 'allowSingleOrDouble',
-          }),
-          createCase(name, {
-            ...test.options,
-            format,
-            trailingUnderscore: 'allowSingleOrDouble',
-          }),
-
-          // prefix
-          createCase(`MyPrefix${name}`, {
-            ...test.options,
-            format,
-            prefix: ['MyPrefix'],
-          }),
-          createCase(`MyPrefix2${name}`, {
-            ...test.options,
-            format,
-            prefix: ['MyPrefix1', 'MyPrefix2'],
-          }),
-
-          // suffix
-          createCase(`${name}MySuffix`, {
-            ...test.options,
-            format,
-            suffix: ['MySuffix'],
-          }),
-          createCase(`${name}MySuffix2`, {
-            ...test.options,
-            format,
-            suffix: ['MySuffix1', 'MySuffix2'],
-          }),
-        );
-      }
-    }
-  }
-
-  return newCases;
-}
-function createInvalidTestCases(
-  cases: Cases,
-): TSESLint.InvalidTestCase<MessageIds, Options>[] {
-  const newCases: TSESLint.InvalidTestCase<MessageIds, Options>[] = [];
-
-  for (const test of cases) {
-    for (const [formatLoose, names] of Object.entries(formatTestNames)) {
-      const format = [formatLoose as PredefinedFormatsString];
-      for (const name of names.invalid) {
-        const createCase = (
-          preparedName: string,
-          options: Selector,
-          messageId: MessageIds,
-          data: Record<string, unknown> = {},
-        ): TSESLint.InvalidTestCase<MessageIds, Options> => {
-          const selectors = Array.isArray(test.options.selector)
-            ? test.options.selector
-            : [test.options.selector];
-          const errorsTemplate = selectors.map(selector => ({
-            messageId,
-            ...(selector !== 'default' &&
-            selector !== 'variableLike' &&
-            selector !== 'memberLike' &&
-            selector !== 'typeLike' &&
-            selector !== 'property' &&
-            selector !== 'method'
-              ? {
-                  data: {
-                    type: selectorTypeToMessageString(selector),
-                    name: preparedName,
-                    ...data,
-                  },
-                }
-              : // meta-types will use the correct selector, so don't assert on data shape
-                {}),
-          }));
-
-          const errors: {
-            data?: { type: string; name: string };
-            messageId: MessageIds;
-          }[] = [];
-          test.code.forEach(() => errors.push(...errorsTemplate));
-
-          return {
-            options: [
-              {
-                ...options,
-                filter: IGNORED_FILTER,
-              },
-            ],
-            code: `// ${JSON.stringify(options)}\n${test.code
-              .map(code => code.replace(REPLACE_REGEX, preparedName))
-              .join('\n')}`,
-            errors: errors,
-          };
-        };
-
-        const prefixSingle = ['MyPrefix'];
-        const prefixMulti = ['MyPrefix1', 'MyPrefix2'];
-        const suffixSingle = ['MySuffix'];
-        const suffixMulti = ['MySuffix1', 'MySuffix2'];
-
-        newCases.push(
-          createCase(
-            name,
-            {
-              ...test.options,
-              format,
-            },
-            'doesNotMatchFormat',
-            { formats: format.join(', ') },
-          ),
-
-          // leadingUnderscore
-          createCase(
-            `_${name}`,
-            {
-              ...test.options,
-              format,
-              leadingUnderscore: 'forbid',
-            },
-            'unexpectedUnderscore',
-            { position: 'leading' },
-          ),
-          createCase(
-            name,
-            {
-              ...test.options,
-              format,
-              leadingUnderscore: 'require',
-            },
-            'missingUnderscore',
-            { position: 'leading', count: 'one' },
-          ),
-          createCase(
-            name,
-            {
-              ...test.options,
-              format,
-              leadingUnderscore: 'requireDouble',
-            },
-            'missingUnderscore',
-            { position: 'leading', count: 'two' },
-          ),
-          createCase(
-            `_${name}`,
-            {
-              ...test.options,
-              format,
-              leadingUnderscore: 'requireDouble',
-            },
-            'missingUnderscore',
-            { position: 'leading', count: 'two' },
-          ),
-
-          // trailingUnderscore
-          createCase(
-            `${name}_`,
-            {
-              ...test.options,
-              format,
-              trailingUnderscore: 'forbid',
-            },
-            'unexpectedUnderscore',
-            { position: 'trailing' },
-          ),
-          createCase(
-            name,
-            {
-              ...test.options,
-              format,
-              trailingUnderscore: 'require',
-            },
-            'missingUnderscore',
-            { position: 'trailing', count: 'one' },
-          ),
-          createCase(
-            name,
-            {
-              ...test.options,
-              format,
-              trailingUnderscore: 'requireDouble',
-            },
-            'missingUnderscore',
-            { position: 'trailing', count: 'two' },
-          ),
-          createCase(
-            `${name}_`,
-            {
-              ...test.options,
-              format,
-              trailingUnderscore: 'requireDouble',
-            },
-            'missingUnderscore',
-            { position: 'trailing', count: 'two' },
-          ),
-
-          // prefix
-          createCase(
-            name,
-            {
-              ...test.options,
-              format,
-              prefix: prefixSingle,
-            },
-            'missingAffix',
-            { position: 'prefix', affixes: prefixSingle.join(', ') },
-          ),
-          createCase(
-            name,
-            {
-              ...test.options,
-              format,
-              prefix: prefixMulti,
-            },
-            'missingAffix',
-            {
-              position: 'prefix',
-              affixes: prefixMulti.join(', '),
-            },
-          ),
-
-          // suffix
-          createCase(
-            name,
-            {
-              ...test.options,
-              format,
-              suffix: suffixSingle,
-            },
-            'missingAffix',
-            { position: 'suffix', affixes: suffixSingle.join(', ') },
-          ),
-          createCase(
-            name,
-            {
-              ...test.options,
-              format,
-              suffix: suffixMulti,
-            },
-            'missingAffix',
-            {
-              position: 'suffix',
-              affixes: suffixMulti.join(', '),
-            },
-          ),
-        );
-      }
-    }
-  }
-
-  return newCases;
-}
-
-const cases: Cases = [
-  // #region default
-  {
-    code: [
-      'const % = 1;',
-      'function % () {}',
-      '(function (%) {});',
-      'class Ignored { constructor(private %) {} }',
-      'const ignored = { % };',
-      'interface Ignored { %: string }',
-      'type Ignored = { %: string }',
-      'class Ignored { private % = 1 }',
-      'class Ignored { #% = 1 }',
-      'class Ignored { constructor(private %) {} }',
-      'class Ignored { #%() {} }',
-      'class Ignored { private %() {} }',
-      'const ignored = { %() {} };',
-      'class Ignored { private get %() {} }',
-      'enum Ignored { % }',
-      'abstract class % {}',
-      'interface % { }',
-      'type % = { };',
-      'enum % {}',
-      'interface Ignored<%> extends Ignored<string> {}',
-    ],
-    options: {
-      selector: 'default',
-      filter: '[iI]gnored',
-    },
-  },
-  // #endregion default
-
-  // #region variable
-  {
-    code: [
-      'const % = 1;',
-      'let % = 1;',
-      'var % = 1;',
-      'const {%} = {ignored: 1};',
-      'const {% = 2} = {ignored: 1};',
-      'const {...%} = {ignored: 1};',
-      'const [%] = [1];',
-      'const [% = 1] = [1];',
-      'const [...%] = [1];',
-    ],
-    options: {
-      selector: 'variable',
-    },
-  },
-  // #endregion variable
-
-  // #region function
-  {
-    code: ['function % () {}', '(function % () {});', 'declare function % ();'],
-    options: {
-      selector: 'function',
-    },
-  },
-  // #endregion function
-
-  // #region parameter
-  {
-    code: [
-      'function ignored(%) {}',
-      '(function (%) {});',
-      'declare function ignored(%);',
-      'function ignored({%}) {}',
-      'function ignored(...%) {}',
-      'function ignored({% = 1}) {}',
-      'function ignored({...%}) {}',
-      'function ignored([%]) {}',
-      'function ignored([% = 1]) {}',
-      'function ignored([...%]) {}',
-    ],
-    options: {
-      selector: 'parameter',
-    },
-  },
-  // #endregion parameter
-
-  // #region property
-  {
-    code: [
-      'class Ignored { private % }',
-      'class Ignored { private "%" = 1 }',
-      'class Ignored { private readonly % = 1 }',
-      'class Ignored { private static % }',
-      'class Ignored { private static readonly % = 1 }',
-      'class Ignored { abstract % = 1 }',
-      'class Ignored { declare % }',
-      'class Ignored { #% }',
-      'class Ignored { static #% }',
-    ],
-    options: {
-      selector: 'classProperty',
-    },
-  },
-  {
-    code: ['const ignored = { % };', 'const ignored = { "%": 1 };'],
-    options: {
-      selector: 'objectLiteralProperty',
-    },
-  },
-  {
-    code: [
-      'interface Ignored { % }',
-      'interface Ignored { "%": string }',
-      'type Ignored = { % }',
-      'type Ignored = { "%": string }',
-    ],
-    options: {
-      selector: 'typeProperty',
-    },
-  },
-  // #endregion property
-
-  // #region parameterProperty
-  {
-    code: [
-      'class Ignored { constructor(private %) {} }',
-      'class Ignored { constructor(readonly %) {} }',
-      'class Ignored { constructor(private readonly %) {} }',
-    ],
-    options: {
-      selector: 'parameterProperty',
-    },
-  },
-  {
-    code: ['class Ignored { constructor(private readonly %) {} }'],
-    options: {
-      selector: 'parameterProperty',
-      modifiers: ['readonly'],
-    },
-  },
-  // #endregion parameterProperty
-
-  // #region method
-  {
-    code: [
-      'class Ignored { private %() {} }',
-      'class Ignored { private "%"() {} }',
-      'class Ignored { private readonly %() {} }',
-      'class Ignored { private static %() {} }',
-      'class Ignored { private static readonly %() {} }',
-      'class Ignored { private % = () => {} }',
-      'class Ignored { abstract %() }',
-      'class Ignored { declare %() }',
-      'class Ignored { #%() }',
-      'class Ignored { static #%() }',
-    ],
-    options: {
-      selector: 'classMethod',
-    },
-  },
-  {
-    code: [
-      'const ignored = { %() {} };',
-      'const ignored = { "%"() {} };',
-      'const ignored = { %: () => {} };',
-    ],
-    options: {
-      selector: 'objectLiteralMethod',
-    },
-  },
-  {
-    code: [
-      'interface Ignored { %(): string }',
-      'interface Ignored { "%"(): string }',
-      'type Ignored = { %(): string }',
-      'type Ignored = { "%"(): string }',
-    ],
-    options: {
-      selector: 'typeMethod',
-    },
-  },
-  // #endregion method
-
-  // #region accessor
-  {
-    code: [
-      'const ignored = { get %() {} };',
-      'const ignored = { set "%"(ignored) {} };',
-      'class Ignored { private get %() {} }',
-      'class Ignored { private set "%"(ignored) {} }',
-      'class Ignored { private static get %() {} }',
-      'class Ignored { static get #%() {} }',
-    ],
-    options: {
-      selector: 'accessor',
-    },
-  },
-  // #endregion accessor
-
-  // #region enumMember
-  {
-    code: ['enum Ignored { % }', 'enum Ignored { "%" }'],
-    options: {
-      selector: 'enumMember',
-    },
-  },
-  // #endregion enumMember
-
-  // #region class
-  {
-    code: ['class % {}', 'abstract class % {}', 'const ignored = class % {}'],
-    options: {
-      selector: 'class',
-    },
-  },
-  // #endregion class
-
-  // #region interface
-  {
-    code: ['interface % {}'],
-    options: {
-      selector: 'interface',
-    },
-  },
-  // #endregion interface
-
-  // #region typeAlias
-  {
-    code: ['type % = {};', 'type % = 1;'],
-    options: {
-      selector: 'typeAlias',
-    },
-  },
-  // #endregion typeAlias
-
-  // #region enum
-  {
-    code: ['enum % {}'],
-    options: {
-      selector: 'enum',
-    },
-  },
-  // #endregion enum
-
-  // #region typeParameter
-  {
-    code: [
-      'class Ignored<%> {}',
-      'function ignored<%>() {}',
-      'type Ignored<%> = { ignored: % };',
-      'interface Ignored<%> extends Ignored<string> {}',
-    ],
-    options: {
-      selector: 'typeParameter',
-    },
-  },
-  // #endregion typeParameter
-];
-
 ruleTester.run('naming-convention', rule, {
   valid: [
-    ...createValidTestCases(cases),
     {
       code: `
         const child_process = require('child_process');
@@ -1496,7 +781,6 @@ ruleTester.run('naming-convention', rule, {
       options: [],
       errors: [{ messageId: 'doesNotMatchFormat' }],
     },
-    ...createInvalidTestCases(cases),
     {
       code: `
         const child_process = require('child_process');
